backend/requirements.txt
fastapi[standard]
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
python-jose[cryptography]>=3.3.0
